streaming/pyproject.toml
[project]
name = "streaming"
version = "0.1.0"
description = "OpenAI SDK Streaming Demo - Learn how to implement real-time streaming with OpenAI API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "asyncio-mqtt>=0.16.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0"
]
